added code to display the book details
